requirements_api.txt
fastapi>=0.115
uvicorn[standard]>=0.30
python-multipart>=0.0.9
python-dotenv>=1.0.1
numpy==1.26.4
scikit-learn==1.4.2
opencv-python-headless>=4.10.0
torch>=2.2.0
